flask-jwt-auth/requirements.txt
bcrypt
blinker
click
dnspython
Flask
Flask-Bcrypt
Flask-JWT-Extended
Flask-PyMongo
itsdangerous
Jinja2
MarkupSafe
PyJWT
pymongo
python-dotenv
Werkzeug
